refactor(super): narrow provision-jobs GET import boundary (#4)
